backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase>=2.0.0,<3.0.0
python-multipart==0.0.6
slowapi==0.1.9
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.0
